requirements.txt
pygame
numpy
pillow
numeric